refactor: move _map_survey_columns into new survey_column_mapping.py
Cuts ColumnMapping, the near-match regexes/helpers, and
_map_survey_columns out of survey.py verbatim into a new sibling
module. Orchestrator now calls
_survey_column_mapping._map_survey_columns(...). Also repoints
tests/test_survey_preview_regressions.py's ColumnMapping import to
the new module, since it imported the class directly from
src.converters.survey.
